fixes: invalid wins, file view commits, sign precedence
A flurry of fixes.

Some invalid wins were encountered during refresh of PR state.

Needed to till the diff view to optinally compare a opened file with the
base commit the branch is based off when opening from aggreaged file
tree. This is what GitHub does and makes it so there's always a valid
diff to look at.

Commnent bubble signs needed to be preferred over "can create comment"
signs

Signed-off-by: ldelossa <louis.delos@gmail.com>