useability: handle rebases
this commit wires in some commit checks when doing a refresh to discern
if a squash or rebase occured.

if it does the repo is reset to its origin via `git reset --hard
{remote}/{branch}`

if the repository is dirty (you have made changes to some code under
review) the refresh will fail and ask you to stash them, keeping any
changes you've made while reviewing in place.

Signed-off-by: ldelossa <louis.delos@gmail.com>