feat: "GHSearchPRs" and "GHRequestedReview"
Introduce two new commands.

"GHRequestedReview" will list all PRs where the user your `gh` tool is
signed in as has been requested to review.

"GHSearchPRs" is a hold over which lists all PRs in the repository. This
is a slow method of doing this until GitHub's search API is used.

Signed-off-by: ldelossa <louis.delos@gmail.com>