feat: "GHNotifications" - implement a notifications buffer
This feature introduces the "GHNotifications" command.

This command will open a new buffer which lists the notifications for
the given repository Neovim is opened to.

From there you can open PRs and issues which the notifications
reference.

By setting a notification as "read" it will be removed from the buffer
view. By setting it to "unsubscribed" it will be removed and any further
non-mention comments will be ignored.