fix: false positives while reading issue/PR numbers
The expression used to match the number was using an asterisk, which
matches zero or more, instead of one or more.

We're also moving away from scanning the line in favor of expanding the
current WORD under the cursor, and extracting the number with lua,
instead of the internal vim functions.