fix: correctly show diffs after thread viewed prior
Closes #46

previous to this commit, if a thread was displayed in a diff view, and
then a new file diff was navigated too, the vimdiff highlighting would
be off.

solution is a hack but simple enough to use, simple close the diff
window and open a new one when rendering a new diff view.

Signed-off-by: ldelossa <louis.delos@gmail.com>