fix: do not require checks to open pr
hackily Closes: #44

for some reason, libuv can sporadically not call the on_read methods for
    particular check_run api calls.

I couldn't find a good solution for this as of right now, but since
checks are additional data not required for pr to be displayed, we'll
just fill them in completely async and not block the pr from opening.

A refresh of the pr will almost always fix any broken check state.

Signed-off-by: ldelossa <louis.delos@gmail.com>