feat,notifications: allow batching of notification actions
this commit adds the ability to select multiple notifications for batch
actions.

this commit allows you to use a keymap to select 1 or more notifications
to mark as read or unsubscribe from in batch.

if no selection is made, the notification actions work as normal, on the
notification under the cursor.

Signed-off-by: ldelossa <louis.delos@gmail.com>