notifications: toggle unread
the notifications API is not 1:1 with the notifications UI on the GitHub
website. Unfortunately, there does not seem to be a way to mark read
messages as "unread". However, we can at least show "all" messages if
you don't see a notification in the default "unread" list.

this commit implements this, along with some helpful useability changes
which attempts to detect if `GHNotifications` can use the current tab
and window or it should create a new one.

Signed-off-by: ldelossa <louis.delos@gmail.com>