feat,commit convo buffer: add a buffer for viewing commit messages and conversations
the GitHub API allows conversations to happen on a per-commit basis.

Previous to this commit there was no way to comment and have a
conversation per commit.

Now, you can access a commit convo buffer but first opening a PR,
navigating to a commit in the PR, and then hitting (default) <CR> on the
root of the commit tree UI that subsequently opened.

This will open a new tab with the full commit message along with the
ability to comment on the commit.

This is also handy for very large commit messages which do not fit
within the detail popup window.