ui: use highlights as message separators
this commit removes the use of utf-8 icons to separate messages and
notifications.

instead, two new highlights "GHThreadSep" and "GHThreadSepAlt" can be
used to define a color which bounds a message and notification.

The latter can be used to alternative colors between messages and
notifications in a thread or list, aiding further in separating the
content visually.