feat,useability: introduce debug buffer for cli interactions
this commit adds a new config flag "debug_logging" and a new command
"GHOpenDebugBuffer".

when `debug_logging` is flipped to `true` in the user config passed to
setup(), all CLI interactions will be outputted into a debug buffer.

this debug buffer can be opened with the command "GHOpenDebugBuffer".
see doc/gh-nvim.txt for more details.