useability: allow opening of prs with deleted branches
when a branch is deleted, usually, the commits are still available
within the remote repository.

this commit allows a pr where the upstream branch has been deleted to be
opened, and explored against the GH API like normal.

a warning message is displayed when opening a pr where the upstream
branch is deleted.

Signed-off-by: ldelossa <louis.delos@gmail.com>