fix: fix review thread conflicts
previous to this commit gh.nvim did not understand that a comment can
track its original commit sha

without this, it could not decipher which comments belonged to which
commits and would overrlay comments over multiple commits.

with these changes will now only display comments on a commit in the
commit tree view, checkout the commit and open the correct files when
using the Conversations: tab, and correctly sync review threads to their
commits during review.

when HEAD is checked out all threads are displayed using their updated
line numbers if the line was moved without change.

when a specific commit is opened the original line numbers will be used
which associate to the original line the comment was made on.

additionally, a few UI elements have changed. the Conversations: tab in
the pr details tree now shows the commit a threaded conversation belongs
to. Also, an asterick is used in the Commits: tree now to indicate which
commit gh.nvim has currently checked out.

fixes: #47

Signed-off-by: Louis DeLosSantos <louis.delos@gmail.com>