fix: do not map comments to commits strictly
@rmacklin has enlightened me on the way GitHub places comments on source
code.

previous to this commit I was under the impression that comments made
specifically at future commits, would not be displayed when browsing
previous commits.

this assumption is false, and GitHub will simply paint a comment on any
file, regardless of the commit the comment was created at, as long as
the file source line matches.

this commit does not perform source file line <> associated comment line
matching yet, this will be explored in an upcoming commit.

Signed-off-by: Louis DeLosSantos <louis.delos@gmail.com>