gitcli: Relax condition of the repo_dirty
When we do git checkout, gh.nvim checks if the repository is "dirty" or
not with gitcli.repo_dirty function. Currently, it checks whether `git
status --porcelain` output is empty or not and if not, displays the
message like "Cannot open pull request because repository has changes.
Stash changes and try again.".

However, this happens when the repo status contains untracked files.
Untracked files cannot be stashed, so this is a bit confusing. Also, it
is generally safe to checkout the repo with untracked files, because if
that checkout overwrites the untracked file, it anyways fails.

This commit relaxes the condition of repo_dirty. If the repo status is
not empty, but only contains untracked files, it considers the repo is
not dirty.

Signed-off-by: Yutaro Hayakawa <yutaro.hayakawa@isovalent.com>